+++ heka/client.py
@@ -188,9 +188,8 @@ class HekaClient(object):
     @property
     def is_active(self):
         # Is this client ready to transmit messages? For now we assume
-        # that if the default sender (i.e. `NoSendSender`) has been
-        # replaced then we're good to go.
-        return not isinstance(self.sender, NoSendSender)
+        # that if a stream is set, we're good to go.
+        return self.stream is not None
 
     def send_message(self, msg):
         # Apply any filters and, if required, pass message along to the