+++ CHANGES.rst
@@ -4,6 +4,7 @@
 - Dropped the entire heka.senders namespace
 - Removed the zeromq dependency as we don't support zeromq in hekad
   yet.
+- Dropped the `NoSendSender`
 
 0.30.1 - 2013-08-16
 ===================